fix: convert numerical to float to avoid json serialization as string

+++ heiplanet_db/postgresql_database.py
@@ -1266,7 +1266,7 @@ def get_var_values_cartesian(
         .all()
     )
     # Convert directly to list of tuples
-    values_list = [(lat, lon, val) for lat, lon, val in values]
+    values_list = [(np.float64(lat), np.float64(lon), val) for lat, lon, val in values]
 
     mydict = {"latitude, longitude, var_value": values_list}
     return mydict
